requirements.txt
coverage>=4.0.3
engarde>=0.3.1
ipython>=4.1.2
jupyter>=1.0.0
matplotlib>=1.5.1
notebook>=4.1.0
numpy>=1.10.4
pandas>=0.17.1
prettyplotlib>=0.1.7
seaborn>=0.7.0
q>=2.6
python-dotenv>=0.5.0
virtualenv>=15.0.1
virtualenvwrapper>=4.7.1
watermark>=1.3.0
pytest>=2.9.2
spotipy>=2.3.8
scikit-learn>=0.17.1
flask